+++ Source/AllOverIt/Cryptography/AES/IAesEncryptorFactory.cs
@@ -1,0 +1,8 @@
+﻿namespace AllOverIt.Cryptography.AES
+{
+    public interface IAesEncryptorFactory
+    {
+        IAesEncryptor Create();
+        IAesEncryptor Create(byte[] key, byte[] iv);
+    }
+}