+++ Source/AllOverIt/Cryptography/Hybrid/IRsaAesHybridEncryptor.cs
@@ -1,0 +1,6 @@
+﻿namespace AllOverIt.Cryptography.Hybrid
+{
+    public interface IRsaAesHybridEncryptor : IEncryptor, IStreamEncryptor
+    {
+    }
+}