+++ Source/AllOverIt/Cryptography/IStreamEncryptor.cs
@@ -1,0 +1,10 @@
+﻿using System.IO;
+
+namespace AllOverIt.Cryptography
+{
+    public interface IStreamEncryptor
+    {
+        void Encrypt(Stream source, Stream destination);
+        void Decrypt(Stream source, Stream destination);
+    }
+}