+++ Source/AllOverIt/Cryptography/RSA/IRSACryptoServiceProviderFactory.cs
@@ -1,9 +1,0 @@
-﻿using System.Security.Cryptography;
-
-namespace AllOverIt.Cryptography.RSA
-{
-    public interface IRSACryptoServiceProviderFactory
-    {
-        RSACryptoServiceProvider Create();
-    }
-}