+++ Source/AllOverIt/Cryptography/RSA/IRSAEncryptor.cs
@@ -2,11 +2,14 @@
 
 namespace AllOverIt.Cryptography.RSA
 {
-    public interface IRSAEncryptor : IEncryptor
+    public interface IRsaEncryptor : IEncryptor
     {
-        int MaxInputLength { get; }
+        IRsaEncryptionConfiguration Configuration { get; }
 
-        byte[] Encrypt(byte[] data, RSAParameters parameters);
-        byte[] Decrypt(byte[] data, RSAParameters parameters);
+        // Derived from the key used for encryption (typically the public key)
+        int GetMaxInputLength();
+
+        byte[] Encrypt(byte[] plainText, RSAParameters parameters);
+        byte[] Decrypt(byte[] cipherText, RSAParameters parameters);
     }
 }