+++ Source/AllOverIt/Cryptography/RSA/IRsaEncryptionConfiguration.cs
@@ -1,0 +1,10 @@
+﻿using System.Security.Cryptography;
+
+namespace AllOverIt.Cryptography.RSA
+{
+    public interface IRsaEncryptionConfiguration
+    {
+        RsaKeyPair Keys { get; }
+        RSAEncryptionPadding Padding { get; }
+    }
+}