+++ Source/AllOverIt/Cryptography/RSA/IRsaFactory.cs
@@ -1,0 +1,13 @@
+﻿using System.Security.Cryptography;
+
+using RSAAlgorithm = System.Security.Cryptography.RSA;
+
+namespace AllOverIt.Cryptography.RSA
+{
+    public interface IRsaFactory
+    {
+        RSAAlgorithm Create();
+        RSAAlgorithm Create(int keySizeInBits);
+        RSAAlgorithm Create(RSAParameters parameters);
+    }
+}