+++ Change Log.md
@@ -35,7 +35,7 @@
 * Added Type extension methods GetEnumerableElementType(), CreateList()
 * Added SelectManyAsync() extension method
 * Added support for executing a Process and not wait for it to complete
-* Added some basic encryption helpers
+* Added RSA, AES and RSA-AES hybrid encryption helpers
 
 
 ## AllOverIt.AspNetCore